Reduce debug buffer by size of internal buffer
Make the debug buffer and internal debug buffer 1kB in total, rather
than 1kB and 64 bytes respectively.

In practice, this only reduces the debug buffer by 64 bytes, but can
save quite a bit due to alignment in linker script. E.g., on HiFive1 it
ends up saving a whole 1kB since we ask for the APP_MEMORY buffer to be
1kB aligned, and without this change the total kernel memory ends up
just 64 bytes over a 1kB boundary.