libraries/tock-cells: remove unsafe from TakeCell
When TakeCell was first written, the Rust standard library for Cell<T>
required the contained type T to implement Copy. To achieve the
desired functionality, UnsafeCell was used instead. The definition of
Cell has since evolved to the point where unsafe code is no longer
required to implement TakeCell, so this commit refactors it out.