register_structs!: Fix breakage of custom_test_frameworks
The current version of the `register_structs!` macro is not usable in `no_std`
libraries or binaries that implement `#![feature(custom_test_frameworks)]`.

Current `register_structs!` definitions will break the test execution because
they emit unit tests with the `#[test]` attribute macro, which depends on the
`std` test crate which isn't there and can't be used in `no_std`.

Fix this by making the `std` unit tests opt-out through a feature.