cortex-m: nvic to new regs interface; dynamic nvic
Previously, we were statically assuming the number of interrupts available on a
board. Either we were getting lucky and all of the chips had the same number of
interrupts, or, (more likely) it's a no-op to enable/disable NVICs that aren't
present, so the core was silently dropping some writes.

Now we poll the hardware to do the right number of NVIC operations.