kernel: UKB: rename and clarify init function
The initialize function for the userland kernel boundary was called
`initialize_new_process()`, but there are times when a process must be
initialized but it is not new, for example when restarting a process.
This renames the function to just `initialize_process()` which is more
clear.

Also, this clarifies that the init function must also init the stored
state for the process.