arch/rv32i: Don't overwrite the return address
Previously when calling switch_to_process() we would overwrite the
return address with the PC we are going to jump to. This doesn't really
make sense and can cause breakages if we interrupt in the middle of
execution.

Don't overwrite the return address of the app from the kernel.

We also make some other changes as reported by make formatall.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>