Add a threat model to Tock's documentation.
This threat model gives a concrete description of the isolation guarantees Tock OS intends to provide. There are a few differences between Tock OS' current implementation and the isolation described in this threat model:

1. `tockloader` does not currently validate the TBF `total_size` field

2. TBF header parsing is not currently robust against hostile headers

3. The current virtualization/multiplexing does not provide the non-starvation property requested by this threat model.

Question for reviewers: I think this should be linked from ../README.md, but it's unclear which heading it fits under. Should it be under "Overview and Design of Tock"? "Tock Implementation"? "Interface Details"?