rust: remove all unused feature pragmas
This commit removes every #![feature] pragma not necessary to make
all boards successfully build. This change is the first step towards
building as much of the Tock tree with a stable Rust compiler as
possible.

This change should otherwise be a no-op, and will be followed up by
changes that remove feature that are currently necessary for building
but not actually necessary for Tock to work.