kernel: stdize process_each and process_each_cap
With the updates to AppId, we no longer need to manually expose the
index of a process in the processes array to get its ID. This makes it
possible to make both versions of process_each in the kernel the same.

Also, this prohibits users of this interface from using the processes
array index as an identifier (which it no longer is).