rv32i: pmp: Fix the app allocation calculator
When calculating a size for apps we need to make sure we increment the
size by 1 as RISC-V PMP addresses aren't inclusive, while Tock's are.

The problem is we were doing this after the PowerOfTwo::ceiling call,
which meant if we were below a power of two allignment to start with we
would waste space. Change the order so we increment the size first, then
ceiling it to a power of 2.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>