kernel: review grant_ptr interface
Change grant_ptr to be a safe function that returns `&mut *mut u8`
instead of `*mut *mut u8`.

The array of grant pointers itself is always guaranteed to exist,
however the grant pointed to may not yet be allocated. This simplifies
the internals of grants and removes many unneeded `unsafe` blocks.