Add a link to Tockloader (suggested by Branden) and de-conflict the paragraphs that discuss ownership of ADC samples.
ADC samples are a fuzzy corner of data ownership: they are measurements of global state that are taken on the behalf of processes. Because they measure global state that is un-muxable, they may be shared between multiple processes. This results in the samples being process data, but for multiple processes. Because these samples can be shared fairly freely, the details are not relevant to the threat model. Instead, I made the examples a bit more precise, focusing on examples where the ownership matters (i.e. the UART).