capsule: process console: fix printout ordering
This avoids a potential issue with a virtual uart for the process
console. Since the console echos characters to the console itself, but
uses `debug!()` for a majority of its printouts, the last echoed newline
can get stuck in the virtual uart queue behind the debug!() output,
causing slightly malformed output in the terminal. This makes sure that
the newline has been echoed before calling any debug!() commands.