capsules: virtual uart: fix flag usage
This fixes the issue where receives issued from callbacks were always
resulting in abort() calls by not replacing the underlying UART receive
buffer until after all of the client callbacks have finished. This
"flag", in conjunction with the `completing_read` bool are what
start_receive uses to determine whether to start or abort a receive.

This also fixes determining the minimum length of a receive by
calculating it in all cases where another receive is needed.