Makefile: Initial support for QEMU automation testing
This add two targets to the top level Makefile:
 - emulation-setup: Clones the latest QEMU from mainline and builds it
   for the RV32 platforms. We use master QEMU as it will soon have
   OpenTitan support.
 - emulation-check: Boot Tock in QEMU using the board's Makefile. We
   then do a simple test to see if we printed the "Entering main loop"
   string.

This is based on what libtock-rs used to do (with the timeout command
and grep). This should be moved across to a cargo test runner which is
what libtock-rs does now. We just need to figure out a way to hook the
test runner up with the board's Makefile (to run the QEMU args) to avoid
code duplication and test what users will run.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>