cortex-m: fix MPU cacheing logic
Not really sure what I was thinking the first time, but this now correctly
tracks the hardware configuration in the hardware structure and the per-process
MPU configuration in the process config structure, and will update the hardware
if either don't match.

This also adds some documentation to these structures to save me from myself
in the future.