build: use cargo check directly on workspace
As this checks only syntax, it's not important to have the target specified
for each board (caveat: the limited arch-specific cfg'd sections, which will
be covered by the build rules, and are few in number). It's *much* faster
(60s -> 4s on my machine) to do a plain check as modules aren't re-checked
with different target flags for each board this way.

This also removes the --release specifier from check rules. We don't do much
to differentiate debug/release at the moment, but it seems likely that release
would only remove code, so let's default to checking the larger surface area
of the debug build.