sched: Add a watchdog
This PR starts to implement a watchdog functionality.
This is not instead of a safe wait (in lots of cases a safe wait would be
more graceful) but is generally to help avoid infinite loops and help
debugging. It is also something that has come up to help make Tock more
reliable.

It is up to Chips (and in the future hopefully boards) to handle the
watchdog interrupt and decided how to act.

Currently we tickle the watch dog when iterating on each process
and at the start of each kernel loop.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>