virtual_uart: set device as inflight before starting an op
Before this CL, a call to MuxUart::do_next_op set self.inflight to the
node that had pending operation *after* calling the handler for the
operation (transmitted_buffer or transmitted_word). As a result, if the
operation was performed synchronously, MuxUart::transmitted_buffer was
called while self.inflight was still None. Which led to not calling
transmitted_buffer for the originating device and not returning the
buffer to it. So, after a single debug!, all subsequent debug! calls
didn't have an output_buffer and kept adding to the ring_buffer w/o
printing.
This CL fixes it by setting self.inflight for MuxUart before performing
the operatiion. Note that we can only set it if we know tx_buffer exists
(otherwise can end up for a non-empty inflight with no operation pending),
so need to set inside the `node.tx_buffer.take().map` block.

BUG=none
TEST=make flash (w/o https://crrev.com/i/2951789 applied); check that
     all debug! output is present.

Change-Id: I1ed6c074d57c25e36e4ba788f7589a46f493da0b