Consolidate creation of descriptor buffers
Really, descriptors.rs should be handling all of this and the control
endpoint handler should just be working with buffers of data. That
wasn't totally accomplished here, but it was a start. Additionally, this
means that descriptors created by the upper layers can just be made on
the stack and passed in with no need for persistant storage.