kernel: tbf parse: fix bad header length error
An app with a too short header length value in the header length field
was causing the TBF parsing to throw an internal error. This re-works
the logic to be much more inline with the Tock Threat Model where we can
trust the total_length field. Now we can be much more specific that only
with certain errors we stop parsing app headers, and in general we skip
over apps where headers cannot be parsed to see if there are additional
apps after the invalid one.

BUG=none
TEST=make

(cherry picked from commit b33334fc6bc7f24a00787ccdec786c324b5acfbd)
Change-Id: I23b1518b4088de1b1c809ad3a82f5a3e5bb7d827