chip: rename Ibex to EarlGrey
Currently EarlGrey is the first and only OpenTitan SoC design, with
Ibex CPU at the heart.

Ibex being a CPU does not have any knowledge of EarlGrey peripherals,
and hence was misrepresented in Tock. This has caused some confusion
to how exactly is was intended to be used, and did not achieve the
goals of being stand alone crate, as it was coupled to heavily with
the OpenTitan board.