kernel: grant: do not pass T::default()
This changes how grant allocation is structured in grant.rs so that the
allocation function does not take a copy of the data to be written when
creating a new grant. This data can be large if the grant region is
large, and it can cause a stack overflow when the allocate function is
called.

(cherry picked from commit 8059404925858c41e006587011361f47d7ceb086)
Change-Id: Idcc0294bf74350950539b1aa14f1a8014180d406