boards: make: remove .apps from kernel binary
The .apps section in flash is a placeholder section that satisfies two
properties:

1. Its physical address is the same address as the start of apps in
flash.
2. The linker places it into a segment, so that when the .elf file is
used, the .apps section will be loaded at the correct address.

In the kernel build system, we have no application binaries, so the
.apps section only includes a placeholder byte so that the linker keeps
it around and creates a segment for it. Later, the .elf file can be
updated with an actual application binary in the .apps section, and the
application binary will be loaded with the kernel binary.

Due to the requirements above, that single placeholder byte means the
kernel .elf file creates a binary with one byte at the address where
applications start. This, when flashed, overwrites the first application
header, effectively erasing all apps.

However, when we are creating just the kernel binary file in the build
system, we do not need to include the .apps section (it is just a
placeholder after all). So, we explicitly ignore it in the .elf file
when creating the kernel binary.