Remove the use of #[feature(const_fn)] from OptionalCell.
tock-cells still needs `const_fn` for TakeCell, as creating types containing mutable references in a const function requires `const_fn`. This is one of two approaches to removing the dependency OptionalCell has on `const_fn`; the other is to use an `impl<C: Copy>` block. I'll send that approach in a separate PR from this commit.