arch/rv32i: Reduce PMP stack usage
This is the first step to reducing the stack usage from the PMP config structs.

Before this patch the stack usage was:

    720     _ZN68_$LT$rv32i..pmp..PMPConfig$u20$as$u20$kernel..platform..mpu..MPU$GT$13configure_mpu17h5967285bd6ec5881E

After this patch it is now:

    592     _ZN68_$LT$rv32i..pmp..PMPConfig$u20$as$u20$kernel..platform..mpu..MPU$GT$13configure_mpu17h5967285bd6ec5881E

The configure_mpu() function is still the second highest user of the
stack (after reset_handler()) but this is an improvement.

This is a step to help fix: https://github.com/tock/tock/pull/2016

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>