tools: allow borrow_interior_mutable_const
This lint triggers when using interior mutability behind a pointer in a
const variable, a common pattern when dealing with memory-mapped
registers. We didn't encounter this error until switching writeable
registers to use `UnsafeCell`.

The error is a false positive because what it's guarding against is
accidentally mutating a temporary value created from a const. But the
`UnsafeCell` we're mutating is behind a `*const` ptr, so we don't
actually create a new copy of it when using the const value - only of
the pointer.

Signed-off-by: David Ross <David.Ross@wdc.com>