arch/rv32i: pmp: Move all of PMP code into a macro
After some discussions it was decided to move the entire PMP code into a
macro and call that from each Chip. This commit does this.

The advantage here is that each Chip only ever has one implementation so
we don't rely on the compiler being smart and optimising unused code
out.

The disadvantage here is that we are now further away from migrating to
const generics.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>