register-interface: Allow expressions in register_bitfields
The goal of this behavior is to allow things like

    register_bitfields![usize,
        pub pmpaddr [
            addr OFFSET(0) NUMBITS(XLEN) []
        ]
    ];

where XLEN is the width of usize. This will make it much easier to
create ISA-word-size-agnostic ports. Without these changes, it would be
necessary to declare each bitfield twice with cfg attributes, even when
only the width of the bitfield differs.

The primary drawback to this feature is that register_bitfields is no
longer hygenic. This is a necessary consequence because otherwise the
new submodules have no way to see the register type, or any constants
used in the bit width. This could have been mitigated by creating a
typedef for the register type and using that typedef in each submodule,
but that would not have solved the problem of using constants in
NUMBITS, etc. Unfortunately, there is already one collision,
necessitating the renaming of x to _x. However, I suspect such
collisions will be rare, since files declaring bitfields tend to declare
little else.

Signed-off-by: Sean Anderson <seanga2@gmail.com>