rv32i: Add support for riscv64 feature
Much of the assembly code has to be duplicated. With the new asm!
macro, the many of the constants (e.g. the *4s and *8s) could have been
factored out. However, there is still no way to choose between ld and lw
(and sd and sw).

Most places just use usize instead of u32.

Signed-off-by: Sean Anderson <seanga2@gmail.com>