rv32i: Add support for riscv64 to CSRs
The CSRs are mostly the same between riscv32 and riscv64. Most of the
time we just need to change the type to Reg and sub in XLEN for
constants of 32. The PMP CSR has been modified to make it easier to
write generic code to modify it.

read_cycle_counter has been updated to prevent overflows.

64-bit support has been done as a feature. This allows chips to
explicitly depend on 64-bit support (or lack of it).

Signed-off-by: Sean Anderson <seanga2@gmail.com>