Fix potential unsoundness in dynamic grants
In the original code, dynamic grant allocation directly returns
`Owned`. This is problematic, as if a capsule stores these Owned values
outside of another app-specific grant, then it could keep it around
after the app has crashed/restarted, and then access deallocated
memory.

This fixes that by instead returning a new `DynamicGrant` type, which
requires an `enter` call to access.

Signed-off-by: David Ross <David.Ross@wdc.com>