Create tock-rubble intermediate interface
The new kernel::hil::rubble is an interface separating the Tock kernel
and the rubble implementation. It has clones of various plain-data
rubble structures for communication with rubble, and traits
representing each of rubble's main components, with trait methods to
call rubble's interface.

This commit fully adds the "kernel<->rubble stack" part of the
interface, changes the rubble capsule to use this new interface, and
creates the new tock_rubble implementing this interface.

This also adds a new tock-rubble-interface<->hardware interface living
in the HIL, but this is less complete - and we still depend on rubble
in the actual hardware implementation, as it uses rubble-nrf5x backing.

There are a few TODOs in the code here, but I'm committing & uploading
for general feedback on the interface.

Signed-off-by: David Ross <David.Ross@wdc.com>