Add full address and type accessors to nrf52 Ficr
This changes the `Ficr::address` function to retrieve the full address
as bytes (as opposed to the first 32bits of a 48bit address), and adds
a new function `Ficr::address_type` to retrieve the address type. I'm
adding these mainly so we can easily use the device's builtin address
as its BLE address.

Signed-off-by: David Ross <David.Ross@wdc.com>