Add TakeCell::get_mut function
This bypasses the Cell nature of take_cell for situations where you
already have access to an `&mut TakeCell`. Similarly to how the
corresponding method in `Cell`, `Cell::get_mut` is safe,
`TakeCell::get_mut` is entirely safe and requires no new unsafe code to
implement. It's made safe by the fact that the returned value is
lifetime bound to the takecell itself, not to `'a`.

This was useful for an odd interaction with the rubble interface when
implementing tock_rubble - it needs to retrieve an `&mut [u8]` multiple
times safely, and then also use that buffer as an `&'static mut [u8]`
which can be taken out. Thus a `TakeCell`, with the addition of this
`get_mut` method, works well.

Signed-off-by: David Ross <David.Ross@wdc.com>