Initial implementations for the STM32f3, STM32f4, msp432, and Apollo3.
Note: I believe there is a bug in the Apollo3 stimer driver but as I
don't have hardware I can't verify it. It sets the compare value
to be a delta, rather than an absolute value. This is what the
errata says to do, but from the datasheet I am wondering if this
is an ambiguity/error in the errata.