Fix a pair of bugs in the virtualizer. Both of them relate to problems when there are long running computations in the kernel that start alarms.
The first bug related to deciding if a new alarm was sooner than the
current earliest alarm. The old logic checked if the current earliest
expiration didn't fall within the new alarm's interval. The bug here
is what happens if the current earliest was set earlier in this
execution of the kernel loop (e.g., at boot, setting up the alarm test)
but by the time the newer alarm is started it has already expired.
So,

|----|         earliest
       |----|  new

in the old logic the new alarm would be seen as earlier. The issue is
not being within the interval of the newer alarm means the existing is
either earlier or later; there's now a test to check that it's later
(by looking at now).

The second bug related to deciding if an alarm had expired. The logic
was using the alarm compare value as "now" rather than the real now.
This ran into similar to problems above: if there was a long computation
path in the kernel, such that the hardware alarm had expired before a
later alarm was entered, then this later alarm would not have the
expiration time within its interval. This use of a virtual "now" based
on expiration time was intended to handle some edge cases which the
rest of the logic handle correctly, so I moved this back to using
the real "now" rather than a virtual "now" from the past.