panic handler: don't panic if DEBUG_WRITER is not set
During development of a new board one might panic prior to a debug
writer component being registered. Especially if no serial console is
working yet, seeing the blinking LEDs in the recognizable pattern is a
good first indicator.

If no debug writer component is set, the panic handler will itself
panic, resulting in a panic loop. The LED pattern will not be
displayed.

This checks whether the DEBUG_WRITER global variable is set and prints
a message that it should be set with `set_debug_writer_wrapper`
instead.

Due to the static mutable debug writer being inherently unsafe, this
also reduces its visibility.