kernel: process: reconfigure mpu on restart
If a process restarts we need to setup the MPU based on the original app
break and not the one the app had when it crashed.

If an app shrank its memory before crashing then not doing this causes
an immediate MPU fault if the kernel puts the stack outside of the app's
old memory region.

If an app grew its memory before crashing then not doing this causes a
security issue since an app can grow its heap, crash, then restart and
have access to memory above its app break address.