tock-cells: add const mut refs feature
New nightlies break on this package, since the Cell in take_cell.rs
contains a `&mut`, and we're trying to construct one of those from
inside a `const` function. Bisection showed that this was introduced by
a relevant-seeming range of commits.

Enabling this feature unbreaks it. See
https://github.com/rust-lang/rust/issues/57349.

Error message:

error[E0723]: mutable references in const fn are unstable
  --> libraries/tock-cells/src/take_cell.rs:21:18
   |
21 |             val: Cell::new(None),
   |                  ^^^^^^^^^^^^^^^
   |
   = note: see issue #57563
<https://github.com/rust-lang/rust/issues/57563> for more information
   = help: add `#![feature(const_fn)]` to the crate attributes to enable

BUG=b:171501502
TEST=`make build RUST_TOOLCHAIN_VERSION=nightly-2020-09-23 &&
      make build RUST_TOOLCHAIN_VERSION=nightly-2020-09-24`

Change-Id: Id55cf010610db6364c93d47dc53025bd91d68b6d