Rename AppId to ProcessId
The AppId struct changes when an application restarts, and thereby
identifies a runtime instantiation of an application binary, or
process (in line with the threat model) instead of the application
binary itself.

This commit is a mostly mechanical change from AppId to ProcessId,
along with some variable and function renames and documentation
updates in the kernel code.