chips: support imxrt GPT clock frequency scaling
i.MX RT GPT clocks can run off of the IPG clock root (default clock),
or the crystal oscillator. Before this commit, the GPTs were using
the IPG clock, which was assuming a default value. However, this
value will change when we update the ARM clock speed, since the IPG
clock is derived from the ARM clock root. We'd like to use the
oscillator as the GPT clock source, since the oscillator frequency
isn't susceptible to clock scaling.

This commit updates the GPT to support either IPG or crystal
oscillator selection. We extend the CCM interface to permit perclk
selection and dividing. We then extend the GPT start method to select
the appropriate clock.

Since the clock frequency now varies at runtime, we remove the 2475MHz
frequency type, which provides a static frequency value. We implement
a dynamic frequency inside the GPT module. Each GPT dispatches to the
dynamic frequency value in calls to hil::timer::Timer.

The commit should be backwards compatible. We patched the imxrt1050
EVK board to continue using the IPG clock frequency. In the teensy4
board, we select the crystal oscillator and a divider of 8. This
results in a 1MHz GPT frequency, perfect for 1us intervals.