Made updates following discussion with Leon on 11/18   1. Removed ReturnValue types from kernel::syscall   2. syscalls to use Result types from kernel::driver   3. added code to flatten Result types to registers on syscall return   4. Added generic Result type to kernel::syscall since it must be able      to pass to architecture method for flattening.   5. Updated system call TRD with ReadOnly/ReadWrite allow.
The major open question for software structure  is #3. The ability to
flatten Result types is something a Driver shouldn't care about, but
it does care about the types. So currently the flattening implementations
are in kernel::driver, but I could imagine their being placed in
kernel::syscall.

The code should now be in a good state for the transition of the Driver
trait name and beginning new Driver implementations.