TRD/Syscall, kernel: refinement of the system call return types
This commit introduces some updates to the syscall reference document
as well as the kernel implementation as discussed with Leon Schuermann
<leon@is.currently.online> on 2020-11-18, namely:

  1. Remove ReturnValue types from kernel::syscall

  2. Syscalls use result types from kernel::driver

  3. Add code to flatten result types to registers on syscall return

  4. Add generic result type to kernel::syscall since it must be able
     to pass to architecture method for flattening.

  5. Update system call TRD with ReadOnly/ReadWrite allow.

The major open question for software structure is #3. The ability to
flatten Result types is something a Driver shouldn't care about, but
it does care about the types. So currently the flattening
implementations are in kernel::driver, but I could imagine their being
placed in kernel::syscall.

The code should now be in a good state for the transition of the Driver
trait name and beginning new Driver implementations.

Original subject:

    Made updates following discussion with Leon on 11/18