boards: update teensy4 README to qualify support
In our first review, we identified that we may want to only support
one physical board from the Tock board crate. I've selected the
4.0 development board, since

- it's a little older than the Teensy 4.1, which should mean that
  it's easier source, and easier to find documentation and guides
- it's cheaper than the Teensy 4.1, so should be more accessible
- it's the board that I use for development and testing

The Teensy 4.1 board is a "bigger" 4.0 board with more features.
There are many commonalities: they can share a linker script, they
share the same chip, and they have the same pinout for all pins 0 -
33. If it works on the Teensy 4.0, it most likely also works on the
Teensy 4.1 board.

The README makes these ideas clear. We guarantee support for the
4.0 board, and we note that the crate may work on the 4.1 board.
We will not accept 4.1-specific features in this board.

We can consider how to support Tock on the 4.1 boards later. Couple
of options could be

- feature flags to conditionally add 4.1-specific features
- a completely separate board crate