Tock 2.0 transition: expose current "legacy" trait along new one
This renames the current "Driver" trait to "LegacyDriver" to make
breaking changes to the interface while keeping a backwards-compatible
interface during porting. "LegacyDriver" should be removed entirely
prior to releasing Tock 2.0.

This also updates the `with_driver` Platform-method to accept both
capsules implementing the new and the old driver interface.