TRD/Syscall: allow processes to read RO-allowed buffers with kernel
The way I read the TRD before this change, a process cannot read data
in a buffer while the kernel has access to that data through the
read-only allow call. In `libtock-rs`, I would like to give read-only
allow a signature similar to:

```rust
fn ro_allow(driver: usize, buffer_id: usize, slice: &[u8]);
```

When user code (a binary that depends on `libtock-rs`) calls the above
`ro_allow` function, the user code may be holding onto another
reference to the buffer pointed to by `slice`. `ro_allow` has no way
to know whether this is the case (either at compile time or
runtime). As a result, it would be easy for user code to violate its
"when userspace shares a buffer, it can no longer access it"
constraint.

We could resolve this in `libtock-rs` by creating an `AllowBuffer`
type that tracks at runtime whether it is shared with the kernel, and
require user code touching such buffers to use `AllowBuffer`, but that
looks like it will be pretty unpleasant. I'd rather change the syscall
design so that `libtock-rs` and user code can directly use `&[u8]` to
refer to read-only buffers shared with the kernel.

Original subject:

    Change the syscall TRD to allow processes to read Read-Only Allow
    buffers while the kernel has access to them.