sched: Move system call handling to separate function
This moves the process system call handling logic out of `do_process`
into a separate function. `do_process` is by itself a long function
with multiple levels of indentation. The Tock 2.0 system call
interface will add further complexity. This change is done to ensure
better readability of the function and to reduce the indentation
levels.