imxrt10xx: Fix GPT2 clock gating, unused lifetime
Previous work added the GPT2 static instance. However, the new
instance was still referring to the GPT1 clock gate. This commit
corrects that bug. We expose the GPT2 clock methods from the CCM,
add a new clock gate enumeration, then correctly instantiate the
GPT instances with the independent clock gates.

The commit also correct for an unused lifetime in the gpt typdefs.