Implement Tock 2.0 read-only/-write allow infrastructure
This adds the basic framework for routing an allow system call from
the scheduler to a driver capsule.

It does so by calling a method on Process which must make sure that
the allowed memory region is valid in the process context and is not
an alias of an already allowed buffer. The call to the Driver is done
using a closure, which provides the new AppSlice as an argument and
returns the old AppSlice in its return value.