Redesign AppSlice in accordance with the Tock 2.0 semantics
This splits up the single AppSlice types into two types, as they
contain different pointer types:
- ReadOnlyAppSlice
- ReadWriteAppSlice

Furthermore, it removes the AppPtr to AppSlice abstraction, along with
the `Private` contraint type and the ability to define AppSlices over
types other than u8.

The old types are kept in a legacy module to keep compatibility with
the current Driver trait as part of the Tock 2.0 transition.