Revise the description of Subscribe's semantics to match its implementation.
The implementation of subscribe in the kernel is unable to guarantee that a callback is not called after a subsequent subscribe call. Instead, it works by passing drivers non-copyable Callback types which drivers can use to invoke callbacks. This updates the TRD to match the implementation.

The new wording highlights the difference between what drivers *can* do and what they are *supposed to* do. This is important for memory safety in the context of Tock's threat model: an app must not encounter undefined behavior if a driver misbehaves, although it is okay for driver misbehavior to result in denial of service.

This change doesn't change the behavior of a "well-behaved" driver; it does require apps to be somewhat tolerant to non-well-behaved drivers.