Suppress clippy warning `not_unsafe_ptr_arg_deref` in allow-code
This clippy lint is triggered by accepting a raw pointers in a safe
public Rust method, which dereferences these or passes them onto an
unsafe Rust method.

In the places where this is done we are either never effectively
dereferencing this pointer down the line, or making sure that it is
within bounds of a userspace process. Therefore, it should be fine to
ignore this lint.