Tock 2.0: Deprecate SuccessWithValue
Removes all uses of ReturnCode::SUccessWithValue outside LegacyDriver::command(),
and updates the porting guide to reflect that part of porting requires
removing all uses of SuccessWithValue when moving to Driver.

Given that SuccessWithValue is no longer a consideration, add a helper
function to the kernel for converting a ReturnCode to a CommandResult.

This change does require slight ABI changes to both the UDP driver and
hd44780 driver, as they rely on functionality that does not exist in
2.0.