Callback: differentiate between default instance and null callback
This changes Callback to differentiate between a default instance and
a null callback, which is important for assuring that a capsule does
not swap Callbacks of two different processes.

This does not yet limit the way default instances of Callbacks can be
created or removes the Copy trait on Callback, which would be required
to ensure that a capsule can never keep hold of an old Callback
instance or prevent a capsule from always returning a default instance
back in the `subscribe` Driver method.