Implement Tock 2.0 subscribe, prepare Callback swap prevention
This commit reimplements the Tock 2.0 subscribe operation in a
seperate method on Process. It also starts introducing checks &
infrastructure for detecting when a capsule / driver swapped Callback
instances illegally.

It is not yet sufficient to enforce that a driver always returns the
previous Callback instance to userspace, as it may instantiate an
arbitrary number of default Callback instances. This also implies that
it may hold onto a number of Callbacks and schedule them even though
they should be unallowed.