rv32i: fix race condition of interrupt in switch_to_process
On rv32i, the machine mode only has a single trap handler. Upon
entering this trap handler, the kernel uses the mscratch CSR to
determine whether the trap came from usermode or kernelmode.

However, in switch_to_process, the mscratch register is written with
the address of the kernel stack _prior_ to disabling interrupts in
machine mode. This introduces a race condition where, if an interrupt
comes in between those two steps, the trap handler determines the trap
to come from usermode where it really occured in kernel mode. As a
consequence, the trap handler saves the CPU state (which belongs at
least in part to the kernel) into the process state.

As soon as the process is scheduled again, the registers of the
kernel (as saved in the trap handler) are unpacked. Furthermore, the
mepc register is set to a program counter that belongs to kernel
mode. The process will immediately fault due to an instruction memory
access violation.

This proposed fix changes the order of writing mscratch and disabling
interrupts. It furthermore switches to using the "Atomic Read and
Clear/Set Bit in CSR" instructions to avoid an interrupt from arriving
and changing mstatus between the CSR read and CSR write.

Signed-off-by: Leon Schuermann <leon@is.currently.online>