Partially revert 50743b25f3 (reserving process identifier 0)
This reverts an artifact of the commit
50743b25f37080936ecd0ffb82140a039347b925, which introduced a change
that made process identifiers as used in AppSlices start from 1 to
reserve 0 as a default value enabling architectural changes w.r.t. the
Tock 2.0 system call API. The current approach is to instead wrap
AppSlice in an Option<>-type, which is None when not referring to an
actual process.