rv32i: print captured mcause in process state dump
This CL prints the mcause value captured in the process status at the
beginning of trap handler. This doesn't add much info for normal case
since RISC-V status is printed before the process status, and it
contains mcause. But in some corner cases it helps to understand what is
going on.

BUG=b:176921367
TEST=trigger exception, check panic output

Change-Id: Icf6c37b4833b970659e9a5541d9bb5f6d60596b4