boards: make: do not page align sections
Previously we told the linker the page size was small so that it would
in effect remove any padding between program sections. This was a hack
when switching linkers, but the proper way to do this is using the
"-nmagic" flag which explicitly tells the linker to not align sections
to page sizes.

Some relevant discussions:
- https://github.com/rust-embedded/cortex-m-quickstart/pull/95#issuecomment-691711983
- https://github.com/tock/tock/pull/2180

Also copying in the linker docs which might be useful:

```
$ ld.lld --help | grep magic
  --nmagic                Do not page align sections, link against static libraries.
  --no-nmagic             Page align sections (default)
  --no-omagic             Do not set the text data sections to be writable, page align sections (default)
  -N                      Alias for --omagic
  -n                      Alias for --nmagic
  --omagic                Set the text and data sections to be readable and writable, do not page align sections, link against static libraries
```

Fixes #2219