[WIP] capsules: take process-bound callbacks on process_init
This changes the capsules to implement the `process_init` method on
Platform, taking a `ProcessCallbackFactory` as an argument and
building process-bound callbacks, replacing the default callbacks of
the initial Grant region.

The process-bound callbacks can then be legally returned on calls to
`subscribe` and help guarantee that a capsule can never swap a
processes' callbacks for ones with other subdriver numbers.

Work in progress: not all capsules have been updated to utilize
process_init for replacing default callbacks with process-bound
callbacks.