rv32i switch_to_process: pass process state struct in specific reg
This assures that the compiler may never choose to pass the pointer to
the Riscv32iStoredState in a register that is otherwise used and
overwritten in the assembly (e.g. t0).

Presumably this is a not a general issue, but specific to this
assembly block, as we first store all registers and later restore
them, but still rely on the values of the registers further down in
the assembly, while overwriting specific ones with temporary values.