Rework the initial stack pointer and memory break the kernel gives to processes.
Previously, the kernel gave the app 3 KiB of usable memory and initialized the stack pointer to the top of that memory. This was a backwards-compatibility measure to support Tock 1.0 process binaries. This commit makes the UserspaceKernelBoundary responsible for picking the initial stack pointer and process break, and reduces the guarantees the kernel provides to userspace.

In particular:

1. The memory break is now set as small as possible. Process initialization should move the break upwards to where it is needed.
2. The stack pointer is set to the memory break on ARM and set to an unspecified values (currently 0) on RISC-V

This has the following benefits over the previous setup:

1. It supports processes that use less than 3KiB of memory.
2. It makes process runtime setup easier on ARM, as the memory break should unconditionally be set before the stack pointer, whereas this previously depended on the process' stack size and desired break.

I am unsure how much stack space ARM needs at startup, so there is currently a placeholder value. I intend to set that value correctly before merging this.