imxrt10xx: Double stack buffer for i.MX RT boards
Board-based peripheral initialization requires at least

  core::mem::size_of::<Imxrt10xxDefaultPeripherals>()

bytes on the stack to initialize the static peripherals. The majority
of the space comes from the GPIO ports and pins. As of this
writing, that struct size is around 2K, which was half of the
previous stack space. We're running out of stack space, so this
commit doubles the stack size.

I didn't notice the issue until we updated the Rust compiler to
2021-01-07. Before that update, the kernel would initialize, and
apps would run. But, a panic would fail. The panic handler is
additionally allocating a ~2K gpio::Ports struct on the stack, which
ends up exceeding the stack size. This is still true as of this
commit.