Add a description of possible future return variant discriminators.
The existing wording of the Tock 2.0 syscalls TRD is (unintentionally?) vague about whether new return variants can be added in future versions of Tock's 32-bit syscall ABI. This commit makes the wording unambiguous: new return variants may be added in the future, having arbitrary 32-bit discriminants. Processes (e.g. `libtock-rs` and its clients) need to be ready to handle unknown return variants, in case they are run on a Tock kernel newer than the version they are compiled for.

Allowing arbitrary 32-bit variants to be added in the future maximizes the ABI's future flexibility. In particular, we *may* choose to extend the ABI to 64-bit architectures in a manner that adds future variants (e.g. "success with usize"). If we decided we do not want that flexibility in the future, we can further restrict the possible return variant discriminators without breaking compatibility with existing processes.